Add integration tests for HTTP and SQL tagging
Invoke the full stack and fix the various inevitable bugs that came to
the surface.

Add options to `RequestTagger.start` to disable SQL or HTTP. This also
makes unit testing a whole lot simpler as we don't have RSpec doubles
lurking around and breaking tests.

Also add options to pass in a custom identifier for HTTP and SQL tags
(defaulting to `X-Request-Id` and `request-id` respectively).