Improve documentation, tie up some loose ends
For now we'll ask the user to include our concern in their base
controller rather than doing something fancy with railties. Maybe this
is fine for the long term.

Update README to reflect changes.

Require `net/http` in `HttpMagic` module - let's not assume `Net::HTTP`
will be loaded for us (caught this when running a server in production
mode using this gem).